change translations_avail
Changed the default translations_avail to only English.
This can irritate some users who think, VLSub is delivered with these languages at default.
Only when the user adds some langs in the locale folder or use a working URL (without SSL !)
in vlsub.lua on line 1081, only than the real existing langs should be visible in the drop-down menu.

Otherwise you click on a displayed lang and wonder why it's not change. That's confusing.